makefile

+++ ft_memset.c
@@ -6,11 +6,11 @@
 /*   By: hramaros <hramaros@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/19 15:13:35 by hramaros          #+#    #+#             */
-/*   Updated: 2024/02/20 14:10:57 by hramaros         ###   ########.fr       */
+/*   Updated: 2024/02/21 11:12:03 by hramaros         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, size_t n)
+void	*ft_memset(void *s, int c, int n)
 {
 	void	*s_cpy;
